HADOOP-2995. Fixes StreamBaseRecordReader's getProgress to return a floating point number. Contributed by Dave Brosius.
git-svn-id: https://svn.apache.org/repos/asf/hadoop/core/trunk@636704 13f79535-47bb-0310-9956-ffa450edef68

+++ src/contrib/streaming/src/java/org/apache/hadoop/streaming/StreamBaseRecordReader.java
@@ -84,7 +84,7 @@ public abstract class StreamBaseRecordReader implements RecordReader<Text, Text>
     if (end_ == start_) {
       return 1.0f;
     } else {
-      return (in_.getPos() - start_) / (end_ - start_);
+      return ((float)(in_.getPos() - start_)) / ((float)(end_ - start_));
     }
   }
